requirements.txt
flask
rpi_ws281x
adafruit-circuitpython-neopixel
